fix testcase for DataFilter
The test cases named:
	"Matching 'syscall' data value of hooked_syscall as int"
	"Matching 'syscall' data value of sys_enter as string"

Assume that the open syscall is supported and has an index of 2.  This
is only true on x86.  ARM64 doesn't implement open (it uses openat) and
other architectures use differnet indexes.  This patch switches the
syscall to openat and uses event.Openat instead of a literal 2.